+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### v0.4
 - changed Wallhaven class to allow sub-classes to override get_endpoint(), which will allow more effective rate-limiting of requests
+- fixed usage of seed in get_search_pages
 
 #### v0.3
 - use the [ratelimit](https://pypi.org/project/ratelimit/) library to limit API requests to 45 per minute, as per [the official docs](https://wallhaven.cc/help/api#limits)